267 - update metadata in one step
