+++ workers/workers/tasks/declarations.py
@@ -138,15 +138,6 @@ def recreate_bundle(celery_task, dataset_id, **kwargs):
     return task_body(celery_task, dataset_id, **kwargs)
 
 
-@app.task(base=WorkflowTask, bind=True, name='update_bundle_checksum',
-          autoretry_for=(Exception,),
-          max_retries=3,
-          default_retry_delay=5)
-def archive_dataset(celery_task, dataset_id, **kwargs):
-    from workers.tasks.update_bundle_checksum import update_bundle_checksum as task_body
-    return task_body(celery_task, dataset_id, **kwargs)
-
-
 @app.task(base=WorkflowTask, bind=True, name='replace_sda_archive',
           autoretry_for=(Exception,),
           max_retries=3,
